WIP: Refactored CMS Structure and Naming Convention
